letsgo/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
django-cors-headers==3.8.0
djangorestframework==3.12.4
idna==3.2
psycopg2==2.9.1
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
urllib3==1.26.6
